+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/trinkets/ExoticCrystals.java
@@ -33,12 +33,12 @@ public class ExoticCrystals extends Trinket {
 	@Override
 	protected int upgradeEnergyCost() {
 		//5 -> 8(13) -> 10(23) -> 12(35)
-		return 6+2*level();
+		return 6+2*(int)level();
 	}
 
 	@Override
 	public String desc() {
-		return Messages.get(this, "desc", (int)(100*consumableExoticChance(buffedLvl())));
+		return Messages.get(this, "desc", (int)(100*consumableExoticChance((int)buffedLvl())));
 	}
 
 	public static float consumableExoticChance(){
