+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/trinkets/MossyClump.java
@@ -39,12 +39,12 @@ public class MossyClump extends Trinket {
 	@Override
 	protected int upgradeEnergyCost() {
 		//5 -> 10(15) -> 20(35) -> 30(65)
-		return 10+10*level();
+		return 10+10*(int)level();
 	}
 
 	@Override
 	public String desc() {
-		return Messages.get(this, "desc", (int)(100*overrideNormalLevelChance(buffedLvl())));
+		return Messages.get(this, "desc", (int)(100*overrideNormalLevelChance((int)buffedLvl())));
 	}
 
 	public static float overrideNormalLevelChance(){
